added generalized Post minigame scene to display next treatment steps
