+++ src/game/scenes/ZebraCatcher.js
@@ -1,14 +1,7 @@
 import Phaser from "phaser";
 import { MyPlayer } from "../components/MyPlayer";
-import { startSpecialistScene } from "./hospital/Hospital";
 import { createHomeButton } from "../components/HomeButton";
 
-const textStyle = {
-    fontSize: "24px",
-    fill: "black",
-    align: "center",
-    fontFamily: "'Press Start 2P'",
-};
 
 const winText = ["After a long process of monitoring Sam's responses to the tests and treatments so far, it seems that his symptoms closely match that of the Fibrocutaneous Nodular Syndrome disease.\n",
  "It is characterized by Firm, raised nodules or plaques on the skin, often slow-growing and painless, flesh-colored or pinkish in appearance, may have a slightly purplish hue, may be mistaken for a benign skin lesion, but can invade deeper tissues if left untreated.",
@@ -30,6 +23,7 @@ export class ZebraCatcher extends Phaser.Scene {
         this.target = null;
         this.points = 0;
         this.textScore = null;
+        this.winText = winText;
     }
 
     create() {
@@ -149,7 +143,10 @@ export class ZebraCatcher extends Phaser.Scene {
             2000,
             () => {
                 message.destroy();
-                this.scene.start("ZebraCatcherEnd");
+                this.scene.start("MinigamePost", {
+                    doctorType: "Dermatologist",
+                    winText: this.winText,
+                });
             },
             [],
             this
@@ -165,63 +162,5 @@ export class ZebraCatcher extends Phaser.Scene {
             this.scene.restart();
         }
     }
-}
-
-export class ZebraCatcherEnd extends Phaser.Scene {
-    constructor() {
-        super("ZebraCatcherEnd");
-    }
-
-    create() {
-        const background = this.add.image(
-            this.cameras.main.width / 2,
-            this.cameras.main.height / 2,
-            "bg"
-        );
-        background.setDisplaySize(
-            this.cameras.main.width,
-            this.cameras.main.height
-        );
-
-        // Title text
-        this.add
-            .text(
-                this.game.config.width / 2,
-                this.game.config.height / 4,
-                "Treatment results",
-                { ...textStyle, fontSize: "48px" }
-            )
-            .setOrigin(0.5);
-        
-
-        const selectButton = this.add.sprite(this.cameras.main.width-64, this.cameras.main.height-64, "select").setScale(0.2).setInteractive();
-        let line = this.displayString(winText[0]);
-        let clickCount = 1;
-        selectButton.on('pointerup', () => {
-            clickCount++;
-            line.destroy()
-            // Display the next string from the array
-            if (clickCount <= winText.length) {
-                line = this.displayString(winText[clickCount - 1]);
-            } else {
-                startSpecialistScene(this, "Dermatologist");
-            }
-        });
-    }
-
-    displayString(string) {
-         return this.add
-            .text(
-                this.game.config.width / 2,
-                this.game.config.height / 2,
-                string,
-                {
-                    ...textStyle,
-                    wordWrap: { width: this.game.config.width - 100 },
-                }
-            )
-            .setOrigin(0.5);
-    }
-
 }
