+++ src/game/scenes/index.js
@@ -10,8 +10,9 @@ import { GameOverTileJump } from "./GameOverTileJump";
 import { Maze } from "./Maze";
 import { Pong, gameEnd } from "./Pong";
 import { iSpy } from "./iSpy";
-import { ZebraCatcher, ZebraCatcherEnd } from "./ZebraCatcher";
+import { ZebraCatcher } from "./ZebraCatcher";
 import { MinigameMenu } from "./MinigameMenu";
+import { MinigamePost } from "./MinigamePost";
 
 // IMPORTANT: you must import and add your new scene to this array for it to be renderable!!
 export const scenes = [
@@ -30,6 +31,6 @@ export const scenes = [
     iSpy,
     ZebraCatcher,
     MinigameMenu,
-    ZebraCatcherEnd,
+    MinigamePost,
 ];
